+++ jni/src/faiss_methods.cpp
@@ -32,12 +32,15 @@ faiss::IndexIDMapTemplate<faiss::IndexBinary>* FaissMethods::indexBinaryIdMap(fa
 void FaissMethods::writeIndex(const faiss::Index* idx, const char* fname) {
     faiss::write_index(idx, fname);
 }
+
 void FaissMethods::writeIndexBinary(const faiss::IndexBinary* idx, const char* fname) {
     faiss::write_index_binary(idx, fname);
 }
+
 faiss::Index* FaissMethods::readIndex(faiss::IOReader* f, int io_flags) {
     return faiss::read_index(f, io_flags);
 }
+
 faiss::IndexBinary* FaissMethods::readIndexBinary(faiss::IOReader* f, int io_flags) {
     return faiss::read_index_binary(f, io_flags);
 }